+++ asset/fetcher.go
@@ -7,8 +7,8 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
-	"net/url"
 	"os"
+	"strings"
 	"time"
 )
 
@@ -20,23 +20,26 @@ var (
 // A Fetcher fetches a file from the specified source and returns an *os.File
 // with the contents of the file found at source.
 type Fetcher interface {
-	Fetch(ctx context.Context, source string) (*os.File, error)
+	Fetch(ctx context.Context, source string, headers map[string]string) (*os.File, error)
 }
 
 // URLGetter gets all content at the specified URL.
-type urlGetter func(context.Context, string) (io.ReadCloser, error)
+type urlGetter func(context.Context, string, map[string]string) (io.ReadCloser, error)
 
 // Get the target URL and return an io.ReadCloser
-func httpGet(ctx context.Context, path string) (io.ReadCloser, error) {
-	parsedURL, err := url.Parse(path)
+func httpGet(ctx context.Context, path string, headers map[string]string) (io.ReadCloser, error) {
+	req, err := http.NewRequest(http.MethodGet, path, nil)
 	if err != nil {
 		return nil, fmt.Errorf("error fetching asset: %s", err)
 	}
+	req = req.WithContext(ctx)
 
-	req := (&http.Request{
-		Method: http.MethodGet,
-		URL:    parsedURL,
-	}).WithContext(ctx)
+	for k, v := range headers {
+		values := strings.Split(v, ",")
+		for _, value := range values {
+			req.Header.Add(k, strings.TrimSpace(value))
+		}
+	}
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -54,12 +57,12 @@ type httpFetcher struct {
 
 // Fetch the file found at the specified url, and return the file or an
 // error indicating why the fetch failed.
-func (h *httpFetcher) Fetch(ctx context.Context, url string) (*os.File, error) {
+func (h *httpFetcher) Fetch(ctx context.Context, url string, headers map[string]string) (*os.File, error) {
 	if h.URLGetter == nil {
 		h.URLGetter = httpGet
 	}
 
-	resp, err := h.URLGetter(ctx, url)
+	resp, err := h.URLGetter(ctx, url, headers)
 	if err != nil {
 		return nil, err
 	}